Update README.md
This is a repository that gathers together multiple components related to water quality collection and reporting. It includes tools for data collection, for quality control, for writing reports, for creating maps related to water quality reports, and stores actual water quality reports that have been written or are in progress.  Reports are in R Markdown and in Quarto.